requirements.txt
itemadapter==0.2.0
pandas==1.2.0
SQLAlchemy==1.3.22
Scrapy==2.4.1
psycopg2==2.8.6
spacy==2.3.5
matplotlib==3.3.3
python-dotenv==0.15.0
